+++ FastAppFramework.Demo/App.xaml.cs
@@ -6,9 +6,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 using FastAppFramework.Core;
+using FastAppFramework.Demo.Models;
+using FastAppFramework.Demo.ViewModels;
+using FastAppFramework.Demo.Views;
 using FastAppFramework.Wpf;
 using Prism.Ioc;
-using Reactive.Bindings;
 
 namespace FastAppFramework.Demo
 {
@@ -17,44 +19,35 @@ namespace FastAppFramework.Demo
     /// </summary>
     public partial class App : FastWpfApplication
     {
+#region Protected Functions
         protected override void OnInitialized()
         {
             base.OnInitialized();
-        }
-        protected override void RegisterTypes(IContainerRegistry containerRegistry)
-        {
-            base.RegisterTypes(containerRegistry);
+
+            var theme = this.Settings.GetValue<ThemeSettings>("theme");
+            if (theme != null)
+            {
+                if (theme.LoadFromFile)
+                    theme.Apply();
+                else
+                    theme.Load();
+            }
         }
         protected override void RegisterNavigationTypes(IContainerRegistry containerRegistry)
         {
-            base.RegisterNavigationTypes(containerRegistry);
-
-            // Register types for navigation.
-            containerRegistry.RegisterForNavigation<HomePage, HomePageViewModel>();
-            containerRegistry.RegisterForNavigation<SamplePage, SamplePageViewModel>();
-            containerRegistry.RegisterForNavigation<DemoSettingsPage, DemoSettingsPageViewModel>();
+            containerRegistry.RegisterForNavigation<FirstWizardFrame, FirstWizardFrameViewModel>();
+            containerRegistry.RegisterForNavigation<LicenseAgreementPage, LicenseAgreementPageViewModel>();
+            containerRegistry.RegisterForNavigation<OverviewPage, OverviewPageViewModel>();
+            containerRegistry.RegisterForNavigation<ThemePage, ThemePageViewModel>();
+            containerRegistry.RegisterForNavigation<TopAppBarPage>();
+            containerRegistry.RegisterForNavigation<SearchableComboBoxPage>();
+            containerRegistry.RegisterForNavigation<DialogPage, DialogPageViewModel>();
+            containerRegistry.RegisterForNavigation<ThemeSettingsPage, ThemeSettingsPageViewModel>();
         }
         protected override void RegisterSettingTypes(IApplicationSettingRegistry settingRegistry)
         {
-            base.RegisterSettingTypes(settingRegistry);
-
-            settingRegistry.Register(new ApplicationSettingInfo<DemoSettings>("demo"){ Variability = Variability.Normal });
+            settingRegistry.Register<ThemeSettings>("theme", new ThemeSettings());
         }
-
-        protected override void RegisterNotifyIconContextMenuItems(ContextMenuContainer container)
-        {
-            base.RegisterNotifyIconContextMenuItems(container);
-
-            container.Add("Clickable Item", new ReactiveCommand().WithSubscribe(() => { MessageBox.Show("'Clickable Item' is clicked!"); }));
-            container.Add("Checkable Item", false);
-            container.Add();
-
-            var subMenu = new ContextMenuContainer();
-            subMenu.Add("Sub Clickable Item", new ReactiveCommand().WithSubscribe(() => { MessageBox.Show("'Sub Clickable Item' is clicked!"); }));
-            subMenu.Add("Sub Checkable Item", false);
-            subMenu.Add("Sub Sub Menu", new ContextMenuItem[]{ new ContextMenuClickItem(){ Title = "Sub Sub Clickable Item", Command = new ReactiveCommand().WithSubscribe(() => { MessageBox.Show("'Sub Sub Clickable Item' is clicked!"); }) }});
-
-            container.Add("Sub Menu", subMenu);
-        }
+#endregion
     }
 }