+++ FastAppFramework.Demo/Views/FirstWizardFrame.xaml.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using MaterialDesignExtensions.Controls;
+
+namespace FastAppFramework.Demo.Views
+{
+    public partial class FirstWizardFrame : UserControl
+    {
+        public FirstWizardFrame()
+        {
+            InitializeComponent();
+        }
+    }
+}