+++ CHANGELOG.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Nothing to do.
+
+## [0.2.0] - 2023-03-12
+### Added
+- CustomPopupComboBox, SearchableComboBox
+- KeyboardManager
+
+### Changed
+- Focusable of SearchBar
 
 ## [0.1.0] - 2023-02-19
 - First release.